backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
click==8.2.1
dnspython==2.7.0
ecdsa==0.19.1
fastapi==0.115.14
h11==0.16.0
idna==3.10
motor==3.7.1
passlib==1.7.4
pyasn1==0.6.1
pydantic==2.11.7
pydantic-core==2.33.2
pymongo==4.13.2
python-dotenv==1.1.1
python-jose==3.5.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
typing-extensions==4.14.1
typing-inspection==0.4.1
uvicorn==0.35.0
